Move SlotStateManager to its own file, making it a regular class instead of an inner class.
